+++ lib/src/Base/Stat/NumericalSampleImplementation.cxx
@@ -1365,7 +1365,7 @@ NumericalPoint NumericalSampleImplementation::computeSkewnessPerComponent() cons
   ReductionFunctor<SkewnessPerComponentPolicy> functor( *this, policy );
   TBB::ParallelReduce( 0, size_, functor );
   NumericalPoint skewness(dimension_);
-  const NumericalScalar factor(size_ * sqrt(size_ - 1) / (size_ - 2));
+  const NumericalScalar factor(size_ * sqrt(size_ - 1.0) / (size_ - 2));
   for (UnsignedInteger i = 0; i < dimension_; ++i) skewness[i] = factor * functor.accumulator_[i + dimension_] / pow(functor.accumulator_[i], 1.5);
   return skewness;
 }
@@ -1446,7 +1446,7 @@ struct CenteredMomentPerComponentPolicy
     for (UnsignedInteger i = 0; i < dimension_; ++i)
     {
       const NumericalScalar val(point[i] - mean_[i]);
-      var[i] += pow(val, k_);
+      var[i] += pow(val, static_cast<int>(k_));
     }
     return var;
   }
