+++ lib/src/Uncertainty/Model/ContinuousDistribution.cxx
@@ -133,7 +133,7 @@ NumericalScalar ContinuousDistribution::computeProbability(const Interval & inte
   NumericalSample nodesAndWeights(getGaussNodesAndWeights());
   // Perform the integration
   const UnsignedInteger marginalNodesNumber(getIntegrationNodesNumber());
-  const UnsignedInteger size(static_cast< UnsignedInteger >(round(pow(marginalNodesNumber, dimension))));
+  const UnsignedInteger size(static_cast< UnsignedInteger >(round(pow(static_cast<double>(marginalNodesNumber), static_cast<int>(dimension)))));
   NumericalScalar probability(0.0);
   Indices indices(dimension, 0);
   for (UnsignedInteger linearIndex = 0; linearIndex < size; ++linearIndex)
